test: add model validation tests for rssi and datetime serialization
- Add tests to verify rssi accepts float values (matching actual MQTT data)
- Add tests for datetime last_update field handling
- Add tests for WebSocketMessage JSON serialization with datetime
- Add TestMQTTNodeData class with realistic payload validation
- Fix rssi type hint in panel_service.py update_panel method

These tests would have caught the rssi int/float mismatch issue earlier.

+++ dashboard/backend/app/panel_service.py
@@ -104,7 +104,7 @@ class PanelService:
         current_out: Optional[float] = None,
         temperature: Optional[float] = None,
         duty_cycle: Optional[float] = None,
-        rssi: Optional[int] = None,
+        rssi: Optional[float] = None,
         energy: Optional[float] = None,
         online: bool = True,
         timestamp: Optional[str] = None,
